api/dev-requirements.txt
Flask-Testing==0.4.2
PyYAML==3.11
astroid==1.3.8
coverage==3.7.1
coveralls==0.5
docopt==0.6.2
logilab-common==1.0.2
nose==1.3.7
pep8==1.6.2
pylint==1.4.4
requests==2.7.0

